+++ src/hooks/ready/registerGameSettings.js
@@ -118,9 +118,29 @@ export default function () {
     choices: itemCompendiums,
   });
 
-  game.settings.register("ddb-importer", "entity-feature-compendium", {
-    name: "ddb-importer.entity-feature-compendium.name",
-    hint: "ddb-importer.entity-feature-compendium.hint",
+  // game.settings.register("ddb-importer", "entity-feature-compendium", {
+  //   name: "ddb-importer.entity-feature-compendium.name",
+  //   hint: "ddb-importer.entity-feature-compendium.hint",
+  //   scope: "world",
+  //   config: true,
+  //   type: String,
+  //   isSelect: true,
+  //   choices: itemCompendiums,
+  // });
+
+  // game.settings.register("ddb-importer", "entity-classes-compendium", {
+  //   name: "ddb-importer.entity-classes-compendium.name",
+  //   hint: "ddb-importer.entity-classes-compendium.hint",
+  //   scope: "world",
+  //   config: true,
+  //   type: String,
+  //   isSelect: true,
+  //   choices: itemCompendiums,
+  // });
+
+  game.settings.register("ddb-importer", "entity-race-compendium", {
+    name: "ddb-importer.entity-race-compendium.name",
+    hint: "ddb-importer.entity-race-compendium.hint",
     scope: "world",
     config: true,
     type: String,