+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# v 0.2.5
+
+* Optional class features for Rangers now handle speed changes for Deft Explorer (Roving 6th level)
+* Races go into their own compendium
+
 # v 0.2.4
 
 * Some selection boxes inappropriately enabled.